+++ src/monkeybread/__init__.py
@@ -1,6 +1,6 @@
 from importlib.metadata import version
 
-from . import calc, stat, plot, util
+from . import calc, plot, stat, util
 
 __all__ = ["calc", "stat", "plot", "util"]