+++ src/monkeybread/calc/__init__.py
@@ -1,6 +1,6 @@
 """Module for calculations."""
-from ._kernel_density import kernel_density
-from ._shortest_distances import shortest_distances
 from ._cell_contact import cell_contact
 from ._cell_transcript_proximity import cell_transcript_proximity
+from ._kernel_density import kernel_density
 from ._neighborhood_profile import neighborhood_profile
+from ._shortest_distances import shortest_distances