+++ src/monkeybread/stat/_shortest_distances.py
@@ -3,6 +3,7 @@ from collections import Counter
 from typing import List, Optional, Tuple, Union
 
 import numpy as np
+import pandas as pd
 from anndata import AnnData
 from sklearn.neighbors import NearestNeighbors
 
@@ -13,8 +14,9 @@ def shortest_distances(
     group1: Union[str, List[str]],
     group2: Union[str, List[str]],
     n_perms: Optional[int] = 100,
-    actual: Optional[np.ndarray] = None,
+    actual: Optional[pd.DataFrame] = None,
     threshold: Optional[float] = None,
+    basis: Optional[str] = "spatial",
 ) -> Union[np.ndarray, Tuple[np.ndarray, float, float]]:
     """Calculates an expected distribution of shortest distances via permutation of labels.
 
@@ -41,6 +43,8 @@ def shortest_distances(
         :func:`monkeybread.calc.shortest_distances`.
     threshold
         A distance threshold to use for significance calculation, in coordinate units.
+    basis
+        Coordinates in `adata.obsm[X_{basis}]` to use. Defaults to `spatial`.
 
     Returns
     -------
@@ -49,8 +53,8 @@ def shortest_distances(
     is the array containing the expected distribution. The second element corresponds to the
     threshold, and the third element is the p-value calculated.
     """
-    # Calculates the number of distances under the threshold
-    p_statistic = lambda x: np.count_nonzero(np.less(x, threshold))
+    # Calculates the number of distances within the threshold
+    calc_test_statistic = lambda x: np.count_nonzero(np.less_equal(x, threshold))
 
     # Converts groups to lists if one group provided
     if type(group1) == str:
@@ -58,33 +62,46 @@ def shortest_distances(
     if type(group2) == str:
         group2 = [group2]
 
-    # Pulls out group1 spatial locations and the remaining data
-    group1_data = adata.obsm["X_spatial"][[g in group1 for g in adata.obs[groupby]]]
-    filtered_data = adata[[g not in group1 for g in adata.obs[groupby]]]
+    # Pulls out group1 spatial locations, group2 spatial locations, and number of cells in group2
+    # The reason we need this information is because of the way we implement this statistical test.
+    # Specifically, instead of actually shuffling all labels, recomputing the test statistic each time,
+    # we instead randomly sample coordinates from non-group1 cells and compute the shortest
+    # distances from group1 cells to these sampled coordinates. This approach is equivalent to "true
+    # shuffling" approach
+    group1_coords = adata.obsm[f"X_{basis}"][[g in group1 for g in adata.obs[groupby]]]
+    non_group1_ad = adata[[g not in group1 for g in adata.obs[groupby]]]
+    group_to_counts = Counter(non_group1_ad.obs[groupby])
+    n_coords_in_group2 = sum([group_to_counts[g] for g in group2])
+    non_group1_coords = list(non_group1_ad.obsm[f"X_{basis}"])
 
-    # Create distances + statistics arrays, number of cells in group2
+    # Create both distances and test statistics arrays, number of cells in group2
     all_distances = []
-    n_coords = sum([Counter(filtered_data.obs[groupby])[g] for g in group2])
-    coords = list(filtered_data.obsm["X_spatial"])
-    statistics = np.zeros(n_perms)
+    statistics_under_perm = np.zeros(n_perms)
+
     for i in range(n_perms):
         # For each permutation, randomly sample from the non-group1 coordinates and run
         # shortest distances calculation
-        random_coords = rand.sample(coords, k=n_coords)
+        random_coords = rand.sample(non_group1_coords, k=n_coords_in_group2)
         nbrs = NearestNeighbors(n_neighbors=1).fit(random_coords)
-        distances, indices = nbrs.kneighbors(group1_data)
+        distances, indices = nbrs.kneighbors(group1_coords)
+
         # Add distances to array
         all_distances.extend(distances.transpose()[0])
         if threshold is not None:
-            # Add number of distances below threshold to statistics array
-            statistics[i] = p_statistic(distances.transpose()[0])
+            # Add number of distances within threshold to statistics array
+            statistics_under_perm[i] = calc_test_statistic(distances.transpose()[0])
+
     all_distances = np.array(all_distances)
+
     if actual is None or threshold is None:
         # If observed and thresholds aren't provided, don't do p-value calculation
         return all_distances
-    # Convert actual distances to floats and calculate number under threshold
-    actual_statistic = p_statistic(np.vectorize(np.float)(actual))
+
+    # Calculate number of distances under threshold
+    actual_statistic = calc_test_statistic(actual["distance"])
+
     # Calculate p value by comparing actual statistic to each of the permutations
     # Add pseudocount to permutation
-    p_val = np.mean([1 if statistic > actual_statistic else 0 for statistic in statistics] + [1])
+    p_val = (np.sum(np.where(statistics_under_perm >= actual_statistic, 1, 0)) + 1) / (n_perms + 1)
+
     return all_distances, threshold, p_val