+++ CHANGELOG.md
@@ -8,6 +8,35 @@ and this project adheres to [Semantic Versioning][].
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
+## [0.4.4] - 2022-11-29
+
+### Added
+
+-   `mb.calc.neighborhood_profile` can take in a number of nearest neighbors to consider instead of radius
+
+### Changed
+
+-   `mb.plot.cell_contact_heatmap` sorts categories
+
+## [0.4.3] - 2022-11-28
+
+### Changed
+
+-   `mb.calc.shortest_distances` now returns a DataFrame for easier indexing
+-   `mb.plot.cell_transcript_proximity` uses predetermined colors if previously added by Scanpy
+
+## [0.4.2] - 2022-11-23
+
+### Fixed
+
+-   Fixed calculation in shortest distance statistical test
+
+## [0.4.1] - 2022-11-22
+
+### Added
+
+-   Basis customization to functions using coordinates
+
 ## [0.4.0] - 2022-11-21
 
 ### Added